separate some files from GenerateTemplate.swift to new groups
