+++ Sources/templar/Errors/TemplarErrors.swift
