+++ Sources/templar/Extension/Extensions.swift
